// src/api/generated/models/KDSOrderItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Kitchen Display System order item
 */
export type KDSOrderItem = {
    menu_item_id: string;
    name: string;
    quantity: number;
    modifiers?: Array<string>;
    special_instructions?: (string | null);
};